pyproject.toml
[project]
name = "milvus-geo-bench"
version = "0.1.0"
description = "Benchmark tool for Milvus geo search functionality"
readme = "README.md"
authors = [
    { name = "zhuwenxing", email = "wenxing.zhu@zilliz.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pymilvus==2.7.0rc29",
    "shapely>=2.0.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "python-dotenv>=1.0.0",
    "pyproj>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
]

[project.scripts]
milvus-geo-bench = "milvus_geo_bench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# --- uv index 配置 ---
[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
explicit = true

[tool.uv.sources]
pymilvus = { index = "testpypi" }

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".env",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",    # Line too long (handled by line-length setting)
    "PLR0913", # Too many arguments
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison
    "B008",    # Do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]  # Allow unused imports and imports not at top
"tests/*" = ["PLR2004"]          # Allow magic values in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]

[tool.ruff.lint.isort]
known-first-party = ["milvus_geo_bench"]
force-sort-within-sections = true
split-on-trailing-comma = true
